Update Merge Image
In this program, I have created a filter which takes two Mat input and offset, to merge low frequency of one image with high frequency of other image. But the any number except 128 does not give good result.
This filter will show the low frequency image when screen is near to the screen and high frequency image when your the screen is far,
For this, it works great with cat and dog face, Harley and bike, etc kind of images.